Rename D2Common functions
