+++ SlashGaming-Diablo-II-API/include/c/game_func/d2common/d2common_get_global_belt_record.h
@@ -43,24 +43,26 @@
  *  work.
  */
 
-#include "../../../../include/c/game_func/d2common/d2common_get_equipment_slot_layout.h"
+#ifndef SGD2MAPI_C_GAME_FUNC_D2COMMON_D2COMMON_GET_GLOBAL_BELT_RECORD_H_
+#define SGD2MAPI_C_GAME_FUNC_D2COMMON_D2COMMON_GET_GLOBAL_BELT_RECORD_H_
 
-#include "../../../../include/cxx/game_func/d2common/d2common_get_equipment_slot_layout.hpp"
+#include "../../game_struct/d2_belt_record.h"
 
-void D2_D2Common_GetEquipmentSlotLayout(
-    unsigned int inventory_record_index,
+#include "../../../dllexport_define.inc"
+
+#ifdef __cplusplus
+extern "C" {
+#endif // __cplusplus
+
+DLLEXPORT void D2_D2Common_GetGlobalBeltRecord(
+    unsigned int belt_record_index,
     unsigned int inventory_arrange_mode,
-    D2_EquipmentLayout* out_equipment_slot_layout,
-    unsigned int equipment_slot_index
-) {
-  auto* actual_out_grid_layout = reinterpret_cast<d2::EquipmentLayout*>(
-      out_equipment_slot_layout
-  );
+    struct D2_BeltRecord* out_belt_record
+);
 
-  d2::d2common::GetEquipmentSlotLayout(
-      inventory_record_index,
-      inventory_arrange_mode,
-      actual_out_grid_layout,
-      equipment_slot_index
-  );
-}
+#ifdef __cplusplus
+} // extern "C"
+#endif // __cplusplus
+
+#include "../../../dllexport_undefine.inc"
+#endif // SGD2MAPI_C_GAME_FUNC_D2COMMON_D2COMMON_GET_GLOBAL_BELT_RECORD_H_
