+++ SlashGaming-Diablo-II-API/include/c/game_func/d2common/d2common_get_global_equipment_slot_layout.h
@@ -43,10 +43,10 @@
  *  work.
  */
 
-#ifndef SGD2MAPI_C_GAME_FUNC_D2COMMON_D2COMMON_GET_BELT_SLOT_POSITION_H_
-#define SGD2MAPI_C_GAME_FUNC_D2COMMON_D2COMMON_GET_BELT_SLOT_POSITION_H_
+#ifndef SGD2MAPI_C_GAME_FUNC_D2COMMON_D2COMMON_GET_GLOBAL_EQUIPMENT_SLOT_LAYOUT_H_
+#define SGD2MAPI_C_GAME_FUNC_D2COMMON_D2COMMON_GET_GLOBAL_EQUIPMENT_SLOT_LAYOUT_H_
 
-#include "../../game_struct/d2_positional_rectangle.h"
+#include "../../game_struct/d2_equipment_layout.h"
 
 #include "../../../dllexport_define.inc"
 
@@ -54,11 +54,11 @@
 extern "C" {
 #endif // __cplusplus
 
-DLLEXPORT void D2_D2Common_GetBeltSlotPosition(
-    unsigned int belt_record_index,
+DLLEXPORT void D2_D2Common_GetGlobalEquipmentSlotLayout(
+    unsigned int inventory_record_index,
     unsigned int inventory_arrange_mode,
-    struct D2_PositionalRectangle* out_belt_slot,
-    unsigned int belt_slot_index
+    struct D2_EquipmentLayout* out_equipment_slot_layout,
+    unsigned int equipment_slot_index
 );
 
 #ifdef __cplusplus
@@ -66,4 +66,4 @@ DLLEXPORT void D2_D2Common_GetBeltSlotPosition(
 #endif // __cplusplus
 
 #include "../../../dllexport_undefine.inc"
-#endif // SGD2MAPI_C_GAME_FUNC_D2COMMON_D2COMMON_GET_BELT_SLOT_POSITION_H_
+#endif // SGD2MAPI_C_GAME_FUNC_D2COMMON_D2COMMON_GET_GLOBAL_EQUIPMENT_SLOT_LAYOUT_H_
