+++ SlashGaming-Diablo-II-API/include/c/game_func/d2common/d2common_get_global_inventory_position.h
@@ -43,22 +43,26 @@
  *  work.
  */
 
-#include "../../../../include/c/game_func/d2common/d2common_get_inventory_grid_layout.h"
+#ifndef SGD2MAPI_C_GAME_FUNC_D2COMMON_D2COMMON_GET_GLOBAL_INVENTORY_POSITION_H_
+#define SGD2MAPI_C_GAME_FUNC_D2COMMON_D2COMMON_GET_GLOBAL_INVENTORY_POSITION_H_
 
-#include "../../../../include/cxx/game_func/d2common/d2common_get_inventory_grid_layout.hpp"
+#include "../../game_struct/d2_positional_rectangle.h"
 
-void D2_D2Common_GetInventoryGridLayout(
+#include "../../../dllexport_define.inc"
+
+#ifdef __cplusplus
+extern "C" {
+#endif // __cplusplus
+
+DLLEXPORT void D2_D2Common_GetGlobalInventoryPosition(
     unsigned int inventory_record_index,
     unsigned int inventory_arrange_mode,
-    D2_GridLayout* out_grid_layout
-) {
-  auto* actual_out_grid_layout = reinterpret_cast<d2::GridLayout*>(
-      out_grid_layout
-  );
+    struct D2_PositionalRectangle* out_position
+);
 
-  d2::d2common::GetInventoryGridLayout(
-      inventory_record_index,
-      inventory_arrange_mode,
-      actual_out_grid_layout
-  );
-}
+#ifdef __cplusplus
+} // extern "C"
+#endif // __cplusplus
+
+#include "../../../dllexport_undefine.inc"
+#endif // SGD2MAPI_C_GAME_FUNC_D2COMMON_D2COMMON_GET_GLOBAL_INVENTORY_POSITION_H_
