+++ SlashGaming-Diablo-II-API/include/c/game_func/d2common_func.h
@@ -46,10 +46,10 @@
 #ifndef SGD2MAPI_C_GAME_FUNC_D2COMMON_FUNC_H_
 #define SGD2MAPI_C_GAME_FUNC_D2COMMON_FUNC_H_
 
-#include "d2common/d2common_get_belt_slot_position.h"
-#include "d2common/d2common_get_belt_type_record.h"
-#include "d2common/d2common_get_equipment_slot_layout.h"
-#include "d2common/d2common_get_inventory_grid_layout.h"
-#include "d2common/d2common_get_inventory_position.h"
+#include "d2common/d2common_get_global_belt_record.h"
+#include "d2common/d2common_get_global_belt_slot_position.h"
+#include "d2common/d2common_get_global_equipment_slot_layout.h"
+#include "d2common/d2common_get_global_inventory_grid_layout.h"
+#include "d2common/d2common_get_global_inventory_position.h"
 
 #endif // SGD2MAPI_C_GAME_FUNC_D2COMMON_FUNC_H_
