+++ SlashGaming-Diablo-II-API/include/cxx/game_func/d2common/d2common_get_global_belt_record.hpp
@@ -43,25 +43,22 @@
  *  work.
  */
 
-#include "../../../../include/c/game_func/d2common/d2common_get_belt_slot_position.h"
+#ifndef SGD2MAPI_CXX_GAME_FUNC_D2COMMON_D2COMMON_GET_GLOBAL_BELT_RECORD_HPP_
+#define SGD2MAPI_CXX_GAME_FUNC_D2COMMON_D2COMMON_GET_GLOBAL_BELT_RECORD_HPP_
 
-#include "../../../../include/c/game_struct/d2_positional_rectangle.h"
-#include "../../../../include/cxx/game_func/d2common/d2common_get_belt_slot_position.hpp"
+#include "../../game_struct/d2_belt_record/d2_belt_record_struct.hpp"
 
-void D2_D2Common_GetBeltSlotPosition(
+#include "../../../dllexport_define.inc"
+
+namespace d2::d2common {
+
+DLLEXPORT void GetGlobalBeltRecord(
     unsigned int belt_record_index,
     unsigned int inventory_arrange_mode,
-    D2_PositionalRectangle* out_belt_slot,
-    unsigned int belt_slot_index
-) {
-  auto* actual_out_belt_slot = reinterpret_cast<d2::PositionalRectangle*>(
-      out_belt_slot
-  );
+    BeltRecord* out_belt_record
+);
 
-  d2::d2common::GetBeltSlotPosition(
-      belt_record_index,
-      inventory_arrange_mode,
-      actual_out_belt_slot,
-      belt_slot_index
-  );
-}
+} // namespace d2::d2common
+
+#include "../../../dllexport_undefine.inc"
+#endif // SGD2MAPI_CXX_GAME_FUNC_D2COMMON_D2COMMON_GET_GLOBAL_BELT_RECORD_HPP_
