+++ SlashGaming-Diablo-II-API/include/cxx/game_func/d2common/d2common_get_global_belt_slot_position.hpp
@@ -43,22 +43,23 @@
  *  work.
  */
 
-#include "../../../../include/c/game_func/d2common/d2common_get_belt_type_record.h"
+#ifndef SGD2MAPI_CXX_GAME_FUNC_D2COMMON_D2COMMON_GET_GLOBAL_BELT_SLOT_POSITION_HPP_
+#define SGD2MAPI_CXX_GAME_FUNC_D2COMMON_D2COMMON_GET_GLOBAL_BELT_SLOT_POSITION_HPP_
 
-#include "../../../../include/cxx/game_func/d2common/d2common_get_belt_type_record.hpp"
+#include "../../game_struct/d2_positional_rectangle/d2_positional_rectangle_struct.hpp"
 
-void D2_D2Common_GetBeltTypeRecord(
+#include "../../../dllexport_define.inc"
+
+namespace d2::d2common {
+
+DLLEXPORT void GetGlobalBeltSlotPosition(
     unsigned int belt_record_index,
     unsigned int inventory_arrange_mode,
-    D2_BeltRecord* out_belt_record
-) {
-  auto* actual_out_belt_record = reinterpret_cast<d2::BeltRecord*>(
-      out_belt_record
-  );
+    PositionalRectangle* out_belt_slot,
+    unsigned int belt_slot_index
+);
 
-  d2::d2common::GetBeltTypeRecord(
-      belt_record_index,
-      inventory_arrange_mode,
-      actual_out_belt_record
-  );
-}
+} // namespace d2::d2common
+
+#include "../../../dllexport_undefine.inc"
+#endif // SGD2MAPI_CXX_GAME_FUNC_D2COMMON_D2COMMON_GET_GLOBAL_BELT_SLOT_POSITION_HPP_
