+++ SlashGaming-Diablo-II-API/src/c/game_func/d2common/c_d2common_get_global_belt_slot_position.cc
@@ -43,23 +43,24 @@
  *  work.
  */
 
-#ifndef SGD2MAPI_CXX_GAME_FUNC_D2COMMON_D2COMMON_GET_EQUIPMENT_SLOT_LAYOUT_HPP_
-#define SGD2MAPI_CXX_GAME_FUNC_D2COMMON_D2COMMON_GET_EQUIPMENT_SLOT_LAYOUT_HPP_
+#include "../../../../include/c/game_func/d2common/d2common_get_global_belt_slot_position.h"
 
-#include "../../game_struct/d2_equipment_layout/d2_equipment_layout_struct.hpp"
+#include "../../../../include/cxx/game_func/d2common/d2common_get_global_belt_slot_position.hpp"
 
-#include "../../../dllexport_define.inc"
-
-namespace d2::d2common {
-
-DLLEXPORT void GetEquipmentSlotLayout(
-    unsigned int inventory_record_index,
+void D2_D2Common_GetGlobalBeltSlotPosition(
+    unsigned int belt_record_index,
     unsigned int inventory_arrange_mode,
-    EquipmentLayout* out_equipment_slot_layout,
-    unsigned int equipment_slot_index
-);
+    D2_PositionalRectangle* out_belt_slot,
+    unsigned int belt_slot_index
+) {
+  auto* actual_out_belt_slot = reinterpret_cast<d2::PositionalRectangle*>(
+      out_belt_slot
+  );
 
-} // namespace d2::d2common
-
-#include "../../../dllexport_undefine.inc"
-#endif // SGD2MAPI_CXX_GAME_FUNC_D2COMMON_D2COMMON_GET_EQUIPMENT_SLOT_LAYOUT_HPP_
+  d2::d2common::GetGlobalBeltSlotPosition(
+      belt_record_index,
+      inventory_arrange_mode,
+      actual_out_belt_slot,
+      belt_slot_index
+  );
+}
