+++ SlashGaming-Diablo-II-API/src/c/game_func/d2common/c_d2common_get_global_equipment_slot_layout.cc
@@ -43,26 +43,24 @@
  *  work.
  */
 
-#ifndef SGD2MAPI_C_GAME_FUNC_D2COMMON_D2COMMON_GET_INVENTORY_POSITION_H_
-#define SGD2MAPI_C_GAME_FUNC_D2COMMON_D2COMMON_GET_INVENTORY_POSITION_H_
+#include "../../../../include/c/game_func/d2common/d2common_get_global_equipment_slot_layout.h"
 
-#include "../../game_struct/d2_positional_rectangle.h"
+#include "../../../../include/cxx/game_func/d2common/d2common_get_global_equipment_slot_layout.hpp"
 
-#include "../../../dllexport_define.inc"
-
-#ifdef __cplusplus
-extern "C" {
-#endif // __cplusplus
-
-DLLEXPORT void D2_D2Common_GetInventoryPosition(
+void D2_D2Common_GetGlobalEquipmentSlotLayout(
     unsigned int inventory_record_index,
     unsigned int inventory_arrange_mode,
-    struct D2_PositionalRectangle* out_position
-);
+    D2_EquipmentLayout* out_equipment_slot_layout,
+    unsigned int equipment_slot_index
+) {
+  auto* actual_out_grid_layout = reinterpret_cast<d2::EquipmentLayout*>(
+      out_equipment_slot_layout
+  );
 
-#ifdef __cplusplus
-} // extern "C"
-#endif // __cplusplus
-
-#include "../../../dllexport_undefine.inc"
-#endif // SGD2MAPI_C_GAME_FUNC_D2COMMON_D2COMMON_GET_INVENTORY_POSITION_H_
+  d2::d2common::GetGlobalEquipmentSlotLayout(
+      inventory_record_index,
+      inventory_arrange_mode,
+      actual_out_grid_layout,
+      equipment_slot_index
+  );
+}
