+++ SlashGaming-Diablo-II-API/src/cxx/game_func/d2common/d2common_get_global_belt_record.cc
@@ -47,7 +47,7 @@
  * Latest supported version: 1.14D
  */
 
-#include "../../../../include/cxx/game_func/d2common/d2common_get_equipment_slot_layout.hpp"
+#include "../../../../include/cxx/game_func/d2common/d2common_get_global_belt_record.hpp"
 
 #include <windows.h>
 #include <cstdint>
@@ -60,11 +60,10 @@ namespace d2::d2common {
 namespace {
 
 __declspec(naked) void __cdecl
-D2Common_GetEquipmentSlotLayout_1_00(
+D2Common_GetGlobalBeltRecord_1_00(
     std::intptr_t func_ptr,
-    std::uint32_t inventory_record_index,
-    EquipmentLayout* out_equipment_slot_layout,
-    std::uint32_t equipment_slot_index
+    std::uint32_t belt_record_index,
+    BeltRecord* out_belt_record
 ) {
   ASM_X86(push ebp);
   ASM_X86(mov ebp, esp);
@@ -73,7 +72,6 @@ D2Common_GetEquipmentSlotLayout_1_00(
   ASM_X86(push ecx);
   ASM_X86(push edx);
 
-  ASM_X86(push dword ptr [ebp + 20]);
   ASM_X86(push dword ptr [ebp + 16]);
   ASM_X86(push dword ptr [ebp + 12]);
   ASM_X86(call dword ptr [ebp + 8]);
@@ -87,12 +85,11 @@ D2Common_GetEquipmentSlotLayout_1_00(
 }
 
 __declspec(naked) void __cdecl
-D2Common_GetEquipmentSlotLayout_1_09D(
+D2Common_GetGlobalBeltRecord_1_09D(
     std::intptr_t func_ptr,
-    std::uint32_t inventory_record_index,
+    std::uint32_t belt_record_index,
     std::uint32_t inventory_arrange_mode,
-    EquipmentLayout* out_equipment_slot_layout,
-    std::uint32_t equipment_slot_index
+    BeltRecord* out_belt_record
 ) {
   ASM_X86(push ebp);
   ASM_X86(mov ebp, esp);
@@ -101,7 +98,6 @@ D2Common_GetEquipmentSlotLayout_1_09D(
   ASM_X86(push ecx);
   ASM_X86(push edx);
 
-  ASM_X86(push dword ptr [ebp + 24]);
   ASM_X86(push dword ptr [ebp + 20]);
   ASM_X86(push dword ptr [ebp + 16]);
   ASM_X86(push dword ptr [ebp + 12]);
@@ -115,7 +111,7 @@ D2Common_GetEquipmentSlotLayout_1_09D(
   ASM_X86(ret);
 }
 
-std::intptr_t D2Common_GetEquipmentSlotLayout() {
+std::intptr_t D2Common_GetGlobalBeltRecord() {
   static std::intptr_t ptr = mapi::GetGameAddress(__func__)
       .raw_address();
 
@@ -124,29 +120,26 @@ std::intptr_t D2Common_GetEquipmentSlotLayout() {
 
 } // namespace
 
-void GetEquipmentSlotLayout(
-    unsigned int inventory_record_index,
+void GetGlobalBeltRecord(
+    unsigned int belt_record_index,
     unsigned int inventory_arrange_mode,
-    EquipmentLayout* out_equipment_slot_layout,
-    unsigned int equipment_slot_index
+    BeltRecord* out_belt_record
 ) {
-  std::intptr_t func_ptr = D2Common_GetEquipmentSlotLayout();
+  std::intptr_t func_ptr = D2Common_GetGlobalBeltRecord();
 
   GameVersion running_game_version = d2::GetRunningGameVersionId();
   if (running_game_version <= GameVersion::k1_06B) {
-    D2Common_GetEquipmentSlotLayout_1_00(
+    D2Common_GetGlobalBeltRecord_1_00(
         func_ptr,
-        inventory_record_index,
-        out_equipment_slot_layout,
-        equipment_slot_index
+        belt_record_index,
+        out_belt_record
     );
   } else {
-    D2Common_GetEquipmentSlotLayout_1_09D(
+    D2Common_GetGlobalBeltRecord_1_09D(
         func_ptr,
-        inventory_record_index,
+        belt_record_index,
         inventory_arrange_mode,
-        out_equipment_slot_layout,
-        equipment_slot_index
+        out_belt_record
     );
   }
 }
