+++ SlashGaming-Diablo-II-API/src/cxx/game_func/d2common/d2common_get_global_inventory_grid_layout.cc
@@ -43,22 +43,105 @@
  *  work.
  */
 
-#include "../../../../include/c/game_func/d2common/d2common_get_inventory_position.h"
+/**
+ * Latest supported version: 1.14D
+ */
 
-#include "../../../../include/cxx/game_func/d2common/d2common_get_inventory_position.hpp"
+#include "../../../../include/cxx/game_func/d2common/d2common_get_global_inventory_grid_layout.hpp"
 
-void D2_D2Common_GetInventoryPosition(
+#include <windows.h>
+#include <cstdint>
+
+#include "../../../asm_x86_macro.h"
+#include "../../../cxx/game_address_table.hpp"
+#include "../../../../include/cxx/game_version.hpp"
+
+namespace d2::d2common {
+namespace {
+
+__declspec(naked) void __cdecl
+D2Common_GetGlobalInventoryGridLayout_1_00(
+    std::intptr_t func_ptr,
+    std::uint32_t inventory_record_index,
+    GridLayout* out_grid_layout
+) {
+  ASM_X86(push ebp);
+  ASM_X86(mov ebp, esp);
+
+  ASM_X86(push eax);
+  ASM_X86(push ecx);
+  ASM_X86(push edx);
+
+  ASM_X86(push dword ptr [ebp + 16]);
+  ASM_X86(push dword ptr [ebp + 12]);
+  ASM_X86(call dword ptr [ebp + 8]);
+
+  ASM_X86(pop edx);
+  ASM_X86(pop ecx);
+  ASM_X86(pop eax);
+
+  ASM_X86(leave);
+  ASM_X86(ret);
+}
+
+__declspec(naked) void __cdecl
+D2Common_GetGlobalInventoryGridLayout_1_09D(
+    std::intptr_t func_ptr,
+    std::uint32_t inventory_record_index,
+    std::uint32_t inventory_arrange_mode,
+    GridLayout* out_grid_layout
+) {
+  ASM_X86(push ebp);
+  ASM_X86(mov ebp, esp);
+
+  ASM_X86(push eax);
+  ASM_X86(push ecx);
+  ASM_X86(push edx);
+
+  ASM_X86(push dword ptr [ebp + 20]);
+  ASM_X86(push dword ptr [ebp + 16]);
+  ASM_X86(push dword ptr [ebp + 12]);
+  ASM_X86(call dword ptr [ebp + 8]);
+
+  ASM_X86(pop edx);
+  ASM_X86(pop ecx);
+  ASM_X86(pop eax);
+
+  ASM_X86(leave);
+  ASM_X86(ret);
+}
+
+std::intptr_t D2Common_GetGlobalInventoryGridLayout() {
+  static std::intptr_t ptr = mapi::GetGameAddress(__func__)
+      .raw_address();
+
+  return ptr;
+}
+
+} // namespace
+
+void GetGlobalInventoryGridLayout(
     unsigned int inventory_record_index,
     unsigned int inventory_arrange_mode,
-    D2_PositionalRectangle* out_position
+    GridLayout* out_grid_layout
 ) {
-  auto* actual_out_position = reinterpret_cast<d2::PositionalRectangle*>(
-      out_position
-  );
+  std::intptr_t func_ptr = D2Common_GetGlobalInventoryGridLayout();
 
-  d2::d2common::GetInventoryPosition(
-      inventory_record_index,
-      inventory_arrange_mode,
-      actual_out_position
-  );
+  GameVersion running_game_version = d2::GetRunningGameVersionId();
+  if (running_game_version <= GameVersion::k1_06B) {
+    D2Common_GetGlobalInventoryGridLayout_1_00(
+        func_ptr,
+        inventory_record_index,
+        out_grid_layout
+    );
+  } else {
+    D2Common_GetGlobalInventoryGridLayout_1_09D(
+        func_ptr,
+        inventory_record_index,
+        inventory_arrange_mode,
+        out_grid_layout
+    );
+  }
 }
+
+} // namespace d2::d2common
